changed the name of the library to lib-rest-client
