+++ lib-rest-client/README.md
@@ -1,0 +1,4 @@
+# lib-rest-client  
+lib-rest-client is a library that acts as a bridge between ODIM and plugins.  
+This library holds the interfaces and functions to establish a REST connection with the plugin.  
+The library contacts the plugin through REST calls and collects back the response and passes it to the user.
