+++ lib-rest-client/go.mod
@@ -1,0 +1,5 @@
+module github.com/ODIM-Project/ODIM/lib-rest-client
+
+go 1.13
+
+require github.com/ODIM-Project/ODIM/lib-utilities v0.0.0-20200925145026-eac0549d2f51
