--- svc-plugin-rest-client/README.md
@@ -1,4 +1,0 @@
-# svc-plugin-rest-client  
-svc-plugin-rest-client is a library that acts as a bridge between ODIMRA and plugins.  
-This library holds the interfaces and functions to establish a REST connection with the plugin.  
-The library contacts the plugin through REST calls and collects back the response and passes it to the user.
