--- svc-plugin-rest-client/go.mod
@@ -1,5 +1,0 @@
-module github.com/ODIM-Project/ODIM/svc-plugin-rest-client
-
-go 1.13
-
-require github.com/ODIM-Project/ODIM/lib-utilities v0.0.0-20200925145026-eac0549d2f51
